fix: refactor for conciseness
change ``namespace catalyst::subcommand { ... }`` to
``catalyst::subcommand::parse`` with trailing return type

+++ src/subcommands/doc/runners/clang_doc.cpp
@@ -14,11 +14,9 @@ template <> std::expected<void, std::string> DerivedRunner<DocEngine::ClangDoc>:
 
     std::string out_dir = config.getString("manifest.tooling.doc.out_dir").value_or("docs/");
 
-    if (opts.clean) {
-        if (std::filesystem::exists(out_dir)) {
-            catalyst::logger.info("Cleaning output directory: {}", out_dir);
-            std::filesystem::remove_all(out_dir);
-        }
+    if (opts.clean && std::filesystem::exists(out_dir)) {
+        catalyst::logger.info("Cleaning output directory: {}", out_dir);
+        std::filesystem::remove_all(out_dir);
     }
 
     std::string compile_commands = "build/compile_commands.json";
@@ -40,11 +38,11 @@ template <> std::expected<void, std::string> DerivedRunner<DocEngine::ClangDoc>:
         if (std::filesystem::exists(index_path)) {
             catalyst::logger.info("Opening {}", index_path);
 #ifdef __APPLE__
-            catalyst::processExec({"open", index_path}, std::filesystem::current_path().string());
+            auto _ = catalyst::processExec({"open", index_path}, std::filesystem::current_path().string());
 #elif __linux__
-            catalyst::processExec({"xdg-open", index_path}, std::filesystem::current_path().string());
+            auto _ = catalyst::processExec({"xdg-open", index_path}, std::filesystem::current_path().string());
 #elif _WIN32
-            catalyst::processExec({"start", index_path}, std::filesystem::current_path().string());
+            auto _ = catalyst::processExec({"start", index_path}, std::filesystem::current_path().string());
 #endif
         } else {
             catalyst::logger.warn("Could not find index.html to open.");
